New spec file by Willem Riede <wrrhdev@riede.org>

+++ rxtx.spec
@@ -1,10 +1,10 @@
 Summary: RXTX 
 Name: rxtx
-Version: 1.5
-Release: 2
+Version: 2.1
+Release: 7pre17
 License: LGPL
 Group: Development/Libraries
-Source: rxtx-%{PACKAGE_VERSION}.tar.gz
+Source: rxtx-%{PACKAGE_VERSION}-%{PACKAGE_RELEASE}.tar.gz
 URL: www.rxtx.org
 Buildroot: /var/tmp/rxtx-root
 
@@ -12,37 +12,36 @@ Buildroot: /var/tmp/rxtx-root
 rxtx is an full implementation of java commapi which aims to support RS232
 IEEE 1284, RS485, I2C and RawIO.  This is a developers release. 
 %prep
-%setup -q
+%setup -q -n rxtx-%{version}-%{release}
 
 %build
 export THREADS_FLAG=native
-export CLASSPATH=.:/usr/local/java/lib/jcl.jar:/usr/local/java/lib/comm.jar:/usr/local/java/lib/BlackBox.jar
+./autogen.sh
 CFLAGS="$RPM_OPT_FLAGS" LDFLAGS=-s ./configure --prefix=/usr
 pwd
-make comm
+make
 
 %install
 rm -rf $RPM_BUILD_ROOT
-mkdir -p $RPM_BUILD_ROOT/usr/local/java/jre/lib/ext $RPM_BUILD_ROOT/usr/local/java/jre/lib/i386
-for n in libI2C.so libRS485.so libSerial.so libParallel.so libRaw.so; do
-    install -m 755 -s `uname -m`-pc-linux-gnu/.libs/$n $RPM_BUILD_ROOT/usr/local/java/jre/lib/i386/$n
-done
-install -m 755 -s comm.jar $RPM_BUILD_ROOT/usr/local/java/jre/lib/ext
-install -m 755 -s comm.jar $RPM_BUILD_ROOT/usr/local/java/jre/lib/ext
-echo "Driver=gnu.io.RXTXCommDriver" > $RPM_BUILD_ROOT/usr/local/java/jre/lib/javax.comm.properties
-%files
+mkdir -p $RPM_BUILD_ROOT$JAVA_HOME/jre/lib/ext $RPM_BUILD_ROOT$JAVA_HOME/jre/lib/i386
+make RXTX_PATH=$RPM_BUILD_ROOT$JAVA_HOME/jre/lib/i386/ JHOME=$RPM_BUILD_ROOT$JAVA_HOME/jre/lib/ext install
+echo "Driver=gnu.io.RXTXCommDriver" > $RPM_BUILD_ROOT$JAVA_HOME/jre/lib/ext/gnu.io.rxtx.properties
+
+find $RPM_BUILD_ROOT/usr -xtype f -print | \
+    sed "s@^$RPM_BUILD_ROOT@@g" > INSTALLED_FILES
+ 
+if [ "$(cat INSTALLED_FILES)X" = "X" ] ; then
+    echo "No files!"
+    exit -1
+fi
+
+%files -f INSTALLED_FILES
 %defattr(-,root,root)
 %doc AUTHORS ChangeLog README RMISecurityManager.html COPYING INSTALL PORTING TODO
-/usr/local/java/jre/lib/ext/comm.jar
-/usr/local/java/jre/lib/i386/libSerial.so
-/usr/local/java/jre/lib/i386/libParallel.so
-/usr/local/java/jre/lib/i386/libI2C.so
-/usr/local/java/jre/lib/i386/libRS485.so
-/usr/local/java/jre/lib/i386/libRaw.so
-/usr/local/java/jre/lib/javax.comm.properties
-
 
 %clean
 rm -rf $RPM_BUILD_ROOT
 
 %changelog
+* Sun Mar 21 2004  Willem Riede <wrrhdev@riede.org>
+- adjust spec file to support rpmbuild by ordinary user in Fedora context.
